add getWalletsCsv method support getting multiple csv
